fix compilation errors on linux

+++ cfb.c
@@ -74,7 +74,7 @@ static void cfb16_fill_rectangle(QEditScreen *s,
             d = dest;
             n = w;
 
-            if (((int)d & 3) != 0) {
+            if (((intptr_t)d & 3) != 0) {
                 ((uint16_t *)(void *)d)[0] = col;
                 d += 2;
                 n--;
